structure: create dedicated files per class.
